solution-ae/tp_3/requirements.txt
# Streamlit
streamlit

# API
fastapi
pydantic
uvicorn
pydantic
requests
python-dotenv==1.0.1

# TODO Install the required dependencies
langchain-google-genai